checkpoint: renames module directory as `toShortfin`
