+++ src/library/ShimmedStabilityAIClient/toShortfin/toShortfinBatchedRequestBody.ts
