+++ src/library/ShimmedStabilityAIClient/toShortfin/toShortfinRequestBody.ts
