+++ src/library/ShimmedStabilityAIClient/toShortfin/toShortfinRequestBodyPrompt.ts
